+++ d60.Cirqus.NLog/NLogLoggerFactory.cs
@@ -36,9 +36,31 @@ namespace d60.Cirqus.NLog
                 _logger.Warn(message, objs);
             }
 
+            public override void Warn(Exception exception, string message, params object[] objs)
+            {
+                _logger.WarnException(SafeFormat(message, objs), exception);
+            }
+
             public override void Error(string message, params object[] objs)
             {
                 _logger.Error(message, objs);
+            }
+
+            public override void Error(Exception exception, string message, params object[] objs)
+            {
+                _logger.ErrorException(SafeFormat(message, objs), exception);
+            }
+
+            string SafeFormat(string message, params object[] objs)
+            {
+                try
+                {
+                    return string.Format(message, objs);
+                }
+                catch (Exception)
+                {
+                    return message;
+                }
             }
         }
     }