+++ d60.Cirqus/Logging/Logger.cs
@@ -1,4 +1,6 @@
-﻿namespace d60.Cirqus.Logging
+﻿using System;
+
+namespace d60.Cirqus.Logging
 {
     public abstract class Logger
     {
@@ -8,8 +10,13 @@
         }
 
         public abstract void Debug(string message, params object[] objs);
+        
         public abstract void Info(string message, params object[] objs);
+        
         public abstract void Warn(string message, params object[] objs);
+        public abstract void Warn(Exception exception, string message, params object[] objs);
+        
         public abstract void Error(string message, params object[] objs);
+        public abstract void Error(Exception exception, string message, params object[] objs);
     }
 }